refactor: take friends' refactor directory out
친구에 있던 refactor 디렉토리 밖으로 빼고
전에 있던 파일들 지움
