+++ front/src/components/FriendsSection/FriendList/Component/Component.style.js
