+++ front/src/components/FriendsSection/FriendList/FriendList.js
@@ -1,16 +1,16 @@
 import React, { useReducer, useMemo } from 'react';
-import Header from 'components/FriendsSection/refactor/FriendList/Header/Header';
-import Component from 'components/FriendsSection/refactor/FriendList/Component/Component';
+import Header from 'components/FriendsSection/FriendList/Header/Header';
+import Component from 'components/FriendsSection/FriendList/Component/Component';
 import Alert from 'components/globalComponents/Alert/Alert';
 import {
   FriendListStyle,
   FriendListScrollStyle,
-} from 'components/FriendsSection/refactor/FriendList/FriendList.style';
+} from 'components/FriendsSection/FriendList/FriendList.style';
 import { GET_FRIENDS } from 'queries/friend';
 import useCursorQuery from 'hooks/useCursorQuery';
 import InfinityScroll from 'components/globalComponents/InfinityScroll/InfinityScroll';
-import FriendModal from 'components/FriendsSection/refactor/FriendList/FriendModal/FriendModal';
-import modalReducer from 'components/FriendsSection/refactor/FriendList/modalReducer';
+import FriendModal from 'components/FriendsSection/FriendList/FriendModal/FriendModal';
+import modalReducer from 'components/FriendsSection/FriendList/modalReducer';
 import useOnlineFriends from 'hooks/Online/useOnlineFriends';
 
 function changeModeReducer(state) {
