+++ front/src/components/FriendsSection/FriendList/FriendList.style.js
