+++ front/src/components/FriendsSection/FriendList/FriendModal/FriendModal.style.js
