+++ front/src/components/FriendsSection/FriendList/Header/Header.style.js
