+++ front/src/components/FriendsSection/FriendList/modalReducer.js
