--- front/src/components/FriendsSection/FriendsList/DoneButton.style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-import ButtonStyle from 'components/globalComponents/Button/Button.style';
-
-const DoneButtonStyle = styled(ButtonStyle)`
-  width: 3rem;
-  height: 1.5rem;
-  margin-left: auto;
-  margin-right: 0;
-  border: 1px solid white;
-`;
-
-export default DoneButtonStyle;
