--- front/src/components/FriendsSection/FriendsList/FriendComponent.style.js
@@ -1,25 +1,0 @@
-import styled from 'styled-components';
-
-const FriendComponentStyle = styled.div`
-  display: flex;
-  width: 7rem;
-  height: 1rem;
-  background-color: ${(props) => props.theme.pink};
-  color: black;
-  padding-top: 0.5rem;
-  padding-left: 1rem;
-  padding-bottom: 1rem;
-  font-size: 17px;
-  align-items: center;
-
-  > span {
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-  }
-  > *:first-child {
-    flex: 1;
-  }
-`;
-
-export default FriendComponentStyle;
