--- front/src/components/FriendsSection/FriendsList/FriendsSetModal/ButtonSection.style.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const ButtonSectionStyle = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 9rem;
-`;
-
-export default ButtonSectionStyle;
