+++ front/src/components/FriendsSection/FriendsSection.style.js
