--- front/src/components/FriendsSection/ListPopUpButton.style.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-import ButtonStyle from '../globalComponents/Button/Button.style';
-
-const ListPopUpButtonStyle = styled(ButtonStyle)`
-  position: fixed;
-  bottom: 3rem;
-  right: 2rem;
-  width: 9rem;
-  height: 3.8rem;
-  font-size: 25px;
-`;
-
-export default ListPopUpButtonStyle;
