--- front/src/components/FriendsSection/refactor/FriendsSection.context.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const FriendsSectionContext = React.createContext();
-
-export default FriendsSectionContext;
